two cat problem

+++ ft_printf/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: yizhang <yizhang@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/11/04 18:25:13 by yizhang       #+#    #+#                 */
-/*   Updated: 2022/11/04 18:25:13 by yizhang       ########   odam.nl         */
+/*   Updated: 2023/03/12 19:56:21 by yizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
